src/requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
pydantic-settings==2.12.0
aiofiles==25.1.0
# langchain-unstructured[local]==0.1.6
langchain==1.1.0
PyMuPDF==1.26.6
langchain-community==0.4.1
motor==3.7.1




